milestones/requirements_mac.txt
numpy>=1.21.0
matplotlib>=3.4.0
tensorflow-macos>=2.8.0
tensorflow-metal>=0.5.0
tensorflow-datasets>=4.0.0
pydot>=1.4.2
scikit-learn>=1.0.0
ipdb>=0.13.13
kagglehub>=0.3.9
graphviz>=0.20.3